fix(clippy): project-wide clippy --fix to clear CI warnings
Resolve all clippy errors blocking CI on `next` and a broader set of
latent warnings revealed by `cargo clippy --workspace --all-targets
-- -D warnings`. The `CI` workflow (build-test → Clippy telegram-cli)
was failing on 9 errors in octo-network; the workspace-wide check
surfaced 30+ additional issues across 5 crates.

Auto-applied (`cargo clippy --fix`):
- sort_by_key instead of sort_by + .cmp (4 sites)
- Default derive instead of manual impl
- div_ceil() for manual_div_ceil
- collapsible_if merge
- cloned_ref_to_slice_refs → std::slice::from_ref
- needless_borrow cleanup
- useless_vec removal
- unwrap_or instead of unwrap_or_else
- redundant_field_names cleanup
- bool_assert_comparison → assert!

Manual refactors:
- type_complexity (3): extracted DomainIndexKey/Value, MockMessageBus
- too_many_arguments (4): #[allow] with justification on envelope
  builders (build_create_group, build_unbind_all, irc_session) and
  the make_intent test helper — args map 1:1 to envelope fields
- large_enum_variant: IrcStream (TLS handshake is in-memory)
- dead_code (5): #[allow] with rationale for reserved frame types,
  liveness-tracking fields, mock types in shared test helpers
- assertions_on_constants (2): const { assert!(..) } blocks
- doc_lazy_continuation: blank /// to separate list sections
- unused_imports (2): removed stale KeyStorage import and unused
  `provider` arg pattern
- mock_adapter.rs / mock_network.rs: file-level #![allow(dead_code)]
  since each test binary links only a subset of helpers

Verified:
- cargo clippy --workspace --all-targets -- -D warnings → clean
- cargo clippy -p octo-cli-meta --features telegram-cli --all-targets
  -- -D warnings → clean (was failing in last push)
- cargo clippy --manifest-path determin/Cargo.toml --all-targets
  --all-features -- -D warnings → clean
- cargo fmt --all -- --check → clean
- cargo test --workspace → all pass

+++ crates/octo-adapter-whatsapp/src/store.rs
@@ -1136,7 +1136,7 @@ impl DeviceStore for StoolapStore {
                     app_version_primary: row.get::<i64>(12).map_err(to_store_err)? as u32,
                     app_version_secondary: row.get::<i64>(13).map_err(to_store_err)? as u32,
                     app_version_tertiary: row.get::<i64>(14).map_err(to_store_err)? as u32,
-                    app_version_last_fetched_ms: row.get::<i64>(15).map_err(to_store_err)? as i64,
+                    app_version_last_fetched_ms: row.get::<i64>(15).map_err(to_store_err)?,
                     edge_routing_info: {
                         let v: String = row.get(16).map_err(to_store_err)?;
                         if v.is_empty() {
